+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Zaid Changelog
 
+## X.X.X - YYYY-MM-DD
+
+- Changed: clarify the task for "nepali-writer" command
+
 ## 1.0.3 - 2026-08-10
 - Added: pr-summary --base flag to override the auto-detected base branch